refactor: Controller 패키지 coyote 로 이동
