+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -2,6 +2,7 @@ package com.techcourse.controller;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
+import org.apache.coyote.controller.AbstractController;
 import org.apache.coyote.file.ResourcesReader;
 import org.apache.coyote.http11.HttpStatusCode;
 import org.apache.coyote.http11.path.Path;
