+++ tomcat/src/main/java/org/apache/coyote/controller/Controller.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package org.apache.coyote.controller;
 
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
