--- tomcat/src/main/java/org/apache/coyote/http11/executor/Executor.java
@@ -1,10 +1,0 @@
-package org.apache.coyote.http11.executor;
-
-import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.response.HttpResponse;
-
-public interface Executor {
-    HttpResponse execute(HttpRequest req);
-
-    boolean isMatch(HttpRequest req);
-}
